requirements.txt
fastapi
uvicorn[standard]
openai
googletrans==3.1.0a0
pydantic
python-dotenv
langchain-google-genai